Refactor ilm_explain lookup
