+++ src/data/elasticsearch.rs
@@ -1,11 +1,13 @@
 mod aliases;
 mod data_streams;
+mod ilm_explain;
 mod indices_settings;
 mod nodes;
 mod version;
 
 pub use aliases::*;
 pub use data_streams::*;
+pub use ilm_explain::*;
 pub use indices_settings::*;
 pub use nodes::*;
 pub use version::*;
